requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-autocomplete-light==3.9.4
django-crispy-forms==1.14.0
django-htmx==1.9.0
django-mathfilters==1.0.0
gunicorn==20.0.4
Pillow==9.0.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==5.2.0
